Refactor progress data aggregation
