+++ index.html
@@ -37,6 +37,9 @@
     <script src="src/js/project-parameters.js"></script>
     <script src="src/js/research-parameters.js"></script>
     <script src="src/js/skills-parameters.js"></script>
+    <script src="src/js/story/mars.js"></script>
+    <script src="src/js/story/titan.js"></script>
+    <script src="src/js/story/callisto.js"></script>
     <script src="src/js/progress-data.js"></script>
 
     <!-- Core Scripts -->
