+++ src/js/story/AGENTS.md
@@ -1,0 +1,10 @@
+# Story summaries
+
+## Mars
+The opening chapters cover H.O.P.E.'s activation on Mars and the early efforts to build a sustainable colony. After Earth's mysterious destruction, Mary and the Mars Terraforming Committee guide the player through establishing resources, buildings and special projects.
+
+## Titan
+These chapters deal with probing Earth's remains and uncovering an alien threat. The player eventually travels to Titan, captures a hostile bioform and learns of an impending second strike. The story projects include Earth probe launches and interrogation of the alien.
+
+## Callisto
+The final arc moves the action to Callisto where Operation Sidestep is built in secret. Players excavate an old thruster site, forge engine cores on multiple worlds and ultimately assemble the Sidestep Engine to divert a future attack while cult unrest grows.
