+++ tests/atmosphericCollectorUnlock.test.js
@@ -1,13 +1,10 @@
-const fs = require('fs');
-const path = require('path');
 const vm = require('vm');
+const loadProgress = require('./loadProgress');
 
 describe('Atmospheric Water Collector unlock trigger', () => {
   test('story trigger exists with condition prerequisite', () => {
-    const code = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'progress-data.js'), 'utf8');
     const ctx = {};
-    vm.createContext(ctx);
-    vm.runInContext(code, ctx);
+    loadProgress(ctx);
     const chapters = ctx.progressData.chapters;
     const ev = chapters.find(c => c.id === 'any.awCollector');
     expect(ev).toBeDefined();
