+++ tests/interrogateAlienProject.test.js
@@ -1,14 +1,15 @@
 const fs = require('fs');
 const path = require('path');
 const vm = require('vm');
+const loadProgress = require('./loadProgress');
 
 describe('Interrogate Alien project', () => {
   test('generated from progress data', () => {
     const paramsCode = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'project-parameters.js'), 'utf8');
-    const progressCode = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'progress-data.js'), 'utf8');
     const ctx = {};
     vm.createContext(ctx);
-    vm.runInContext(paramsCode + progressCode + '; this.projectParameters = projectParameters;', ctx);
+    vm.runInContext(paramsCode + '; this.projectParameters = projectParameters;', ctx);
+    loadProgress(ctx);
     const project = ctx.projectParameters.interrogate_alien;
     expect(project).toBeDefined();
     expect(project.repeatable).toBe(true);
