+++ tests/interrogateAlienStoryStepsOnce.test.js
@@ -1,13 +1,13 @@
 const fs = require('fs');
 const path = require('path');
 const vm = require('vm');
+const loadProgress = require('./loadProgress');
 const EffectableEntity = require('../src/js/effectable-entity.js');
 
 describe('Interrogate Alien story steps', () => {
   test('final step only printed once', () => {
     const paramsCode = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'project-parameters.js'), 'utf8');
     const projectCode = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'projects.js'), 'utf8');
-    const progressCode = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'progress-data.js'), 'utf8');
 
     const ctx = {
       console,
@@ -20,7 +20,9 @@ describe('Interrogate Alien story steps', () => {
       colonies: {}
     };
     vm.createContext(ctx);
-    vm.runInContext(paramsCode + projectCode + progressCode + '; this.Project = Project; this.projectParameters = projectParameters;', ctx);
+    vm.runInContext(paramsCode + projectCode, ctx);
+    loadProgress(ctx);
+    vm.runInContext('; this.Project = Project; this.projectParameters = projectParameters;', ctx);
 
     const config = ctx.projectParameters.interrogate_alien;
     const project = new ctx.Project(config, 'interrogate_alien');
