+++ tests/waterPumpUnlockChapter.test.js
@@ -1,13 +1,10 @@
-const fs = require('fs');
-const path = require('path');
 const vm = require('vm');
+const loadProgress = require('./loadProgress');
 
 describe('water pump unlock chapter', () => {
   test('water pump is unlocked at liquid water milestone chapter', () => {
-    const code = fs.readFileSync(path.join(__dirname, '..', 'src/js', 'progress-data.js'), 'utf8');
     const ctx = {};
-    vm.createContext(ctx);
-    vm.runInContext(code, ctx);
+    loadProgress(ctx);
     const chapters = ctx.progressData.chapters;
     const ch1_17 = chapters.find(c => c.id === 'chapter1.17');
     const ch3_2 = chapters.find(c => c.id === 'chapter3.2');
